+++ src/refactoring/components/admin/index.ts
@@ -1,2 +1,3 @@
 export * from './AdminNewProductForm';
 export * from './AdminEditingProduct';
+export * from './AdminCoupon';
