+++ src/refactoring/hooks/index.ts
@@ -2,3 +2,5 @@ export * from './useCart.ts';
 export * from './useCoupon.ts';
 export * from './useProduct.ts';
 export * from './useAdminProduct.ts';
+export * from './useAdminEditingProduct.ts';
+export * from './useAdminCoupon.ts';
